Added last_modified Caching
This type of caching affects the loading of the listings page (we could
do the users page as well if we ever get that functionality. If we
reload the pages for listings multiple times, it will return a 304 Not
Modified if the listing hasn't been modified since the last access.